+++ www/js/core/pluginRegistry.js
@@ -15,6 +15,7 @@ import PluginLinkEditor from '../plugins/PluginLinkEditor.js';
 import PluginDependencies from '../plugins/PluginDependencies.js';
 import PluginEventsPlugin from '../plugins/PluginEvents.js';
 import PluginXYBoard from '../plugins/PluginXYBoard.js';
+import PluginPortfolio from '../plugins/PluginPortfolio.js';
 
 const PluginRegistry = {
   'sample-plugin': SamplePlugin,
@@ -30,6 +31,7 @@ const PluginRegistry = {
   'plugin-dependencies': PluginDependencies,
   'plugin-events': PluginEventsPlugin,
   'plugin-xy-board': PluginXYBoard,
+  'plugin-portfolio-board': PluginPortfolio,
 };
 
 export default PluginRegistry;